Issue 96 test cart counter (#104)
Test cart counter and refactorings
